flipselector: Remove has some nbsp's and random whitespaces
Flipselector has some random whitespaces that do not belong.
Additionally there are a few hidden non-blocking spaces that prevent
edje_cc to compile the default.edc on arm. Appearantly on x86/amd64 this
works normally.

This patch works around T2285  while the issue should remain open as the
real issue, edje_cc treating nbsp as a whitespace on arm, is not fixed
through this patch.

ref T2285

Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>