Emile: Reduce ERR logs on invalid JPEG images
When reading the head of a file, we may get the error that it is
not a JPEG image (which is normal), so we should not print any ERR.

The JPEG header read function can indeed be called to test whether a
file can be opened by the JPEG loader or not (any file).

Note that JPEG files don't have reliable magic numbers, so we
don't check them, but rely on libjpeg instead.

Fixes T2290