bindings/eo: Only install Eo.hh once to avoid failing double install
Eo.hh is listed in both dist_installed_eocxxmainheaders_DATA and
dist_installed_eocxxheaders_DATA this leads to a situation where they file
should be installed but already exists and thus the installation fails.

00:32:27.919 /usr/bin/install: cannot create regular file '/srv/build/workspace/changely_efl_configure_xcb/destdir/include/eo-cxx-1/Eo.hh': File exists
00:32:27.923 Makefile:32620: recipe for target 'install-dist_installed_eocxxmainheadersDATA' failed
00:32:27.923 make[4]: *** [install-dist_installed_eocxxmainheadersDATA] Error

Given that it is the only item in dist_installed_eocxxmainheaders_DATA I suspect
it was but into the other rule by accident.