elm_module: Load not installed modules from build dir with ELM_RUN_IN_TREE
If we want to run from the local build dir without make install before
elm_module would fail to load the modules as it is expecting them in the
final install location.

We can now change this by setting ELM_RUN_IN_TREE to force it load the not
yet installed modules.

ref T2028