autotools: add --disable-libeeze
When cross-compiling, we only want to build edje_cc, embryo_cc
and eet binaries for the host before starting the build for the
target.

This patch allows to disable libeeze in order to shorten the
build time but most of all remove the dependency on libudev.

In normal case it's not recommended hence a warning.

Signed-off-by: Romain Naour <romain.naour@openwide.fr>
Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>