assh: add libassh following the same pattern as liblz4.
libassh is a library that provide a ssh implementation designed to be
integrated into a UI toolkit. It is based on a state machine with callbacks.
This will be the basis of an Ecore_Con_Ssh that can be used for shared
configuration for a user or make terminology a ssh client.