ecore_con: fix double free issue in ecore_con_eet.
Summary:
While creating ecore_con_eet_server, if client is not set to false, then while deleting, it is checking for client and it is by default not false. So set it.
ecore_con_reply object is being deleted twice. So removed code to delete the reply object as it would be deleted after the callback.

Signed-off-by: Srivardhan Hebbar <sri.hebbar@samsung.com>

Reviewers: cedric

Subscribers: cedric

Maniphest Tasks: T2372

Differential Revision: https://phab.enlightenment.org/D2445

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>