Eo: Take eo out of beta.
This is following a last review and a discussion on IRC. Eo has been
stable (apart of a decision a few months ago to support more compilers)
for a long while. Developers are already using it for a while, and it's
stupid to break it for them anyway, so we might as well make this
promise now. There are no plans to change it anymore, and it's been
heavily used and tested throughout the EFL for a few releases now.

I'm tagging it as a feature, although it's not, I'm doing it for the
automatic changelog generation. :)

@feature.