Eo base: Fix eo_constructor's declaration.
Remove it from the constructors section.
This was wrong. This place is for functions that are allowed to be
passed to eo_add() and should be used by bindings to create
constructors.
This is wrong for both cases, as this should always be called, not
optional.

Also remove the redundant legacy: null.