SlideShow: Fix double invocation of ItemClass delete function
When closing the application, the delete function given by the user in
the Item Class was called twice. During the clear process, the function
was called for every built item and then during the item destruction.

This patch fixes this behaviour by calling the user function only during
the item destruction.

Another fix has been added. It zero'es the built list pointer inside the
item. Without this fix, the item points to a list node that is already
freed and can lead during widget deletion to an invalid access inside
the list.

@fix T2394