Gengrid: Updated gengrid to select the last item.
Summary:
Updated gengrid to move item selection in 'vertical' mode to the last
item, when currently selected item is not in the last row. As per the
original behaviour, on moving down from such an item, it takes the
focus out of the gengrid widget. Now it is updated to go to the last
item (in last row), and then on further pressing down, focus comes
out of gengrid.

This is something similar to what happens in other UI systems like in
Ubuntu and Windows, when pressing down on (n-1)th row, the last item
in the nth row gets focused.

Signed-off-by: Vaibhav Gupta <g.vaibhav1@samsung.com>

Reviewers: Hermet, SanghyeonLee, raster

Reviewed By: raster

Subscribers: sachin.dev, singh.amitesh

Differential Revision: https://phab.enlightenment.org/D2418