elm list - theme - make compress moder work again
someone.. somewhere along the way.. likely a long time ago.. broke
compress mode. likely the addition of the ellipsis value as an
ellipsis position or something related to that broke compatibility,
as this used to work. be that as it may... finding out just where is
virtually impossible, so fix it the simpler way - add elipsis values