Revert "theme: return false if elm_object_style_set() failed to set requested style."
This reverts commit 76004dfbec84664e253babc5bf576398a5901395.

We need to change other code also to accommodate this change.
_elm_theme_set should return an enum which tells what failed.
enum {
THEME_APPLY_FAILED,
THEME_DEFAULT_SUCCESS.
THEME_APPLY_SUCCESS
};

Based on that, we decide what needs to be done.
The above code will break the layout theme if incorrect theme are passed.
It should be backported to Elm 1.14.

@fix