Elm.Colorselector: Fix crash in infinite callback loop
In most cases the code would not crash (I can't understand why),
but in some cases Eo would complain about call stack overflow.
Yep, that's a lot of recursive calls.

The mouse_down callback was trying to cancel its triggering
event and then send again the same mouse_down event. Why? I guess
because it was easier that also having a callback for move (drag
the colorselector arrows). Ugly...

This widget is in dire need for some love.