Entry: Fix behaviour of markup filters when setting entry text.
The issue was introduced in 28e6878f7cce36efae1438b5e032eb9c0878016a.
While this is a fix in the sense that it reverts to the previously
defined behaviour (defined both in docs and code), the previous
behaviour was stupid and inconsistent. So now we are back to stupid and
inconsistent.

I can't fix this further, because unfortunately, the fix requires
intrusive changes that are just too dangerous to make this late in the
release cycle. Luckily for us, the issues only surface in a very
specific case that is quite uncommon. In order to trigger the bug you
need to set a very long piece of text on an entry that has text filters
set.

Fixes T2561

@fix