Revert "Eolian: fix eolian helper for windows."
This reverts commit 60b473d949c2582695a0ed7e67dc2e2c32eeee1b.

It breaks our mingw cross build setup we are using on jenkins. In summary it
silently fails to use eolian_gen.exe while it should use the linux native
eolian_gen and thus does not generate the c and h files from the eo ones.

See the ml for details or look at the error here:
https://build.enlightenment.org/job/changely_efl_mingw_x86_64/2115/console