multibuttonentry: fix bug with uneditable MBE use case.
Summary:
If there is a uneditable MBE, the MBE doesn't have to get a entry inside of the box.
It breaks MBE box layout, resize logic.

And check the box focus, entry visible is not enough to cover MBE is uneditable case.
If MBE is uneditable, box_resize callback called when item added. in this case,
we need to check what view type user want. if its not shrink view here, do not make
MBE to shrink mode.

@fix

Test Plan: Run elementary_test -> execute multibuttonentry  sample with editable set as false.

Reviewers: Hermet, seoz, woohyun, cedric

Reviewed By: cedric

Differential Revision: https://phab.enlightenment.org/D2824

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>