elm_hoversel: add elm_hoversel_scroll_enabled_set/get() API
Summary:
Currently, hoversel shows all items at the same time, but to limit
the size of contents and make it scrollable is neccesary to some users.
elm_hoversel_scroll_enabled_set/get() API makes the contents of hoversel
scrollable or not.

Test Plan:
The newly added scroll function is enabled for the second hoversel
in "hoversel" sample of elementary_test.

Reviewers: Hermet, raster, herdsman, cedric, tasn

Subscribers: JackDanielZ, cedric, Hermet

Differential Revision: https://phab.enlightenment.org/D2603

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>