Theme: Scroller.edc-Improvement in left/right/top/bottom arrow behavior.
Summary:
When the scroller size changes by dragging the
corners, the calculations will miss the condition "x+w == x1 + w1".
This is because of format specific computer calculations. So,
improvisation is made in comparisons.

@ fix

Signed-off-by: Umesh Tanwar <umesh.tanwar@samsung.com>

Test Plan:
elementary_test -> scroller -> scroll to the extreme right (right arrow brightness changes).
Now, drag the lower right corner to increase the width of scroller. Drag the corner slowly to see the issue clearly.
The arrow brightness changes alternatively, instead the hbar is positioned to the extreme right.

Reviewers: Hermet, cedric, raster

Subscribers: singh.amitesh, eagleeye, SanghyeonLee, sachin.dev

Differential Revision: https://phab.enlightenment.org/D2796