eo - destruction - ensure child is removed from parent child list
this follows on from cbc1a217bfc8b5c6dd94f0448f19245c43eb05e0 as this
code was correct, but was then causing bugs due to children staying in
their parent lists. this should never have happened and this is really
bad. this fixes this and ensures children on destruction are gone from
their parent lists.

@fix