efl - jp2k loader - guard against openjpeg bug tha causes an abort
so... if you load a non-jp2k file using openjpeg, you can get an abort
deep inside the openjpeg library that we can't do anything about. we
set all error handlers but literally the openjpeg code has ab assert
there that causes this bug. it shouldn't and newer opengjpeg libs have
it removed, but 1.5.2 has it and this causes an untrappable crash.
this is simply bad behavior in openjpeg not allowing it to be used
safely to loade image files. the relevant backtrace:

    w=w@entry=0x7fffffffb548, h=h@entry=0x7fffffffb54c,
    alpha=alpha@entry=0x7fffffffb556 "", map=map@entry=0x7fff29ac2000,
    length=<optimized out>, error=error@entry=0x7fffffffb5bc,
    opts=<optimized out>)
    at modules/evas/image_loaders/jp2k/evas_image_load_jp2k.c:111

the relevant code in openjpeg:

int cio_numbytesleft(opj_cio_t *cio) {
  assert((cio->end - cio->bp) >= 0);
        return cio->end - cio->bp;
}

so that assert is triggered. and nothing can be done about it which is
pretty poor.

so an upgrade of openjpeg should fix this as in newer versions have
dropped the assert line in that function, but until poeople have that from
their distro, this adds magic number checks for file headers that avoids
using openjpeg if it's not "apparently" a jp2k file. this does not
stop a corrupt file or a maliciously designed file still causing this
problem, but it does just result in an abort() and isnn't seemingly an
overflow isse that can be exploted, so if you still suffer, find a way to
upgrade openjpeg to 2.x. until then... this reduces inadvertent damage.

@fix