edje: remove un-used edje_container code
The code have been ifdef'ed out for a along time. I checked back until
the merge of our libs before the 1.8 release. Maybe even before that.

If you have a use case for this code fell free to revert but also enable
the code and make sure it works when doing so.