elm map - fix module load craziness loading all elm modules
so elm map loads every module it can find in a recursive dir walk of
all elm modules. this si nuts. this can accidentallly load OLD modules
and thats a recipe for disaster. so check module arch dir aagainst
module arch string to load the right version and be quiet if module
doesn thave the right symbols - it's the wrong module type.

@fix