ecore_evas_extn: Fix server rendering after restart
Summary:
The server can render, only after the server get the OP_SHOW from the client.
However, if the server relaunches while client is running, the server cannot get the OP_SHOW.
In this case, the client should send the OP_SHOW, when the server is added.

Test Plan: Relaunch a server especially the indicator, while client is running.

Reviewers: raster, cedric, Hermet, woohyun, jaehwan, jypark, jpeg

Subscribers: cedric, seoz

Differential Revision: https://phab.enlightenment.org/D3028

Signed-off-by: Jean-Philippe Andre <jp.andre@samsung.com>