elm - when focus goes back in to a window dont have scrollers jump
there is an annoyance in elm that when focus goes back toa window it
restores focus to the last focused widget. if that widget happens to
be invisible in a scroller then this is really annoying as the
scrollers reset. this makes restoring optional. only if there is a
hilight object AND it's visible - then restore, otherwise don't and
revert focus to the window itself again.

@fix