Eo: add a regression test for broken callback comparison.
This bug was fixed in d889da6b12e6123073edb54d766c9af1fc6db822, but it was too
late at night to start hacking on extracting a regression test and adding it to
the suite, so adding it now.

This test makes sure that we only fallback to string comparison with legacy
events.